Refactored the whole background of Elev8Folio
