+++ src/App.css
@@ -146,24 +146,30 @@
 /* Home Page CSS Start*/
 
 .HomePageHolder {
-  color: rgb(42, 41, 41);
   display: flex;
   height: 100vh;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  background-color: rgb(215, 236, 237);
+  background-color: rgb(254, 242, 215);
 }
 
 .HomePageHeader {
-  background-color: rgb(71, 217, 212);
+
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  border-radius: 25%;
+
   animation: pulse infinite;
-  --animate-duration: 4s;
+  --animate-duration: 8s;
+}
+
+.FolioBg{
+  display: block;
+  height: 50vh;
+  border-radius: 5%;
+
 }
 
 .HomePageTitle {
@@ -180,7 +186,6 @@
   margin-bottom: 2em;
   color: rgb(72, 68, 68);
   display: block;
-  background-color: rgb(71, 217, 212);
 
 }
