+++ src/App.js
@@ -4,6 +4,7 @@ import SumeNav from "./Components/SumeNav";
 import Sume from "./Components/Sume";
 import { useState } from "react";
 import Homepage from "./Components/Homepage";
+import Step1 from "./Components/Step1";
 import "animate.css";
 
 function App() {
@@ -16,7 +17,10 @@ function App() {
   return (
     <>
       {HP ? (
-        <Homepage></Homepage>
+        <>
+          <Homepage></Homepage>
+          <Step1></Step1>
+        </>
       ) : (
         <>
           <Header />
